requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
discord==2.3.2
discord.py==2.3.2
frozenlist==1.4.1
idna==3.6
multidict==6.0.4
python-dotenv==1.0.1
yarl==1.9.4
